+++ tests/unit.test.ts
@@ -85,4 +85,36 @@ describe("Arithmetics", () => {
 		expect(fn({ y: 2 }, [ 1, 2, 3 ])).toBe("(1, 4, 3)");
 		expect(fn({ x: 1, y: 2, z: 3 }, [ 1, 2, 3 ])).toBe("(2, 4, 6)");
 	});
+	it("Performs the scalar multiplication of the vector", () => {
+		const fn = (input1: Input, input2: number) => vector(input1).mul(input2).toString();
+
+		expect(fn([ 1 ], 3)).toBe("(3, 0, 0)");
+		expect(fn([ 1, 2 ], 3)).toBe("(3, 6, 0)");
+		expect(fn([ 1, 2, 3 ], 3)).toBe("(3, 6, 9)");
+		expect(fn({ x: 1, y: 2 }, 3)).toBe("(3, 6, 0)");
+		expect(fn({ x: 1, y: 2 }, 3)).toBe("(3, 6, 0)");
+		expect(fn({ x: 1, y: 2, z: 3 }, 3)).toBe("(3, 6, 9)");
+		expect(fn([ 1 ], -3)).toBe("(-3, 0, 0)");
+		expect(fn([ 1, 2 ], -3)).toBe("(-3, -6, 0)");
+		expect(fn([ 1, 2, 3 ], -3)).toBe("(-3, -6, -9)");
+		expect(fn({ x: 1, y: 2 }, -3)).toBe("(-3, -6, 0)");
+		expect(fn({ x: 1, y: 2 }, -3)).toBe("(-3, -6, 0)");
+		expect(fn({ x: 1, y: 2, z: 3 }, -3)).toBe("(-3, -6, -9)");
+		expect(fn([ 1 ], 0.5)).toBe("(0.5, 0, 0)");
+		expect(fn([ 1, 2 ], 0.5)).toBe("(0.5, 1, 0)");
+		expect(fn([ 1, 2, 3 ], 0.5)).toBe("(0.5, 1, 1.5)");
+		expect(fn({ x: 1, y: 2 }, 0.5)).toBe("(0.5, 1, 0)");
+		expect(fn({ x: 1, y: 2 }, 0.5)).toBe("(0.5, 1, 0)");
+		expect(fn({ x: 1, y: 2, z: 3 }, 0.5)).toBe("(0.5, 1, 1.5)");
+	});
+	it("Performs the vector multiplication", () => {
+		const fn = (input1: Input, input2: Input) => vector(input1).mul(input2).toString();
+
+		expect(fn([ 2 ], [ 3 ])).toBe("(6, 0, 0)");
+		expect(fn([ 1, 2 ], [ 3, 4 ])).toBe("(3, 8, 0)");
+		expect(fn([ 1, 2, 3 ], [ 4, 5, 6 ])).toBe("(4, 10, 18)");
+		expect(fn({ x: 2 }, { x: 3 })).toBe("(6, 0, 0)");
+		expect(fn({ x: 1, y: 2 }, { x: 3, y: 4 })).toBe("(3, 8, 0)");
+		expect(fn({ x: 1, y: 2, z: 3 }, { x: 4, y: 5, z: 6 })).toBe("(4, 10, 18)");
+	});
 });